// data/src/data/sfq/type.ts
import { type Static, Type } from '@sinclair/typebox'

const SFQHeader = {
  id: Type.Optional(
    Type.String({
      description: 'The ID of the SFQ report, generated by the database.',
      maxLength: 64,
    }),
  ),
  termCode: Type.String({
    description: 'The academic term.',
    examples: ['FALL', 'SPRING', 'WINTER', 'SUMMER'],
  }),
  termNumber: Type.Number({
    description: 'The term number, starting from 0 for Fall 2011.',
    multipleOf: 1,
  }),
  courseGroup: Type.String({
    description: 'The group of the course.',
    examples: ['COMP'],
  }),
  courseNumber: Type.String({
    description: 'The number of the course.',
    examples: ['2012'],
  }),
  responseRate: Type.Number({
    description: 'The response rate of the report.',
  }),
  surveyName: Type.String({
    description: 'The name (subject) of the survey.',
    default: '',
  }),
} as const

export const SFQCourseReport = Type.Object({
  ...SFQHeader,
  levelOfStatistics: Type.Literal('Course'),
  courseOverallMean: Type.Number({
    description: 'The overall mean of ratings of the course.',
  }),
})

export const SFQInstructorReport = Type.Object({
  ...SFQHeader,
  levelOfStatistics: Type.Literal('Instructor'),
  enrollment: Type.Number(),
  instructorName: Type.Optional(Type.String({
    description: 'The name of the instructor.',
  })),
  courseOverallMean: Type.Number({
    description: 'The overall mean of ratings of the course.',
  }),
  instructorOverallMean: Type.Number({
    description: 'The overall mean of ratings of the instructor.',
  }),
})

export type SFQCourseReport = Static<typeof SFQCourseReport>
export type SFQInstructorReport = Static<typeof SFQInstructorReport>
export type SFQReport = SFQCourseReport | SFQInstructorReport

export const InstructorLevelOfStatistics
  = SFQInstructorReport.properties.levelOfStatistics.const